+++ .gitignore
@@ -5,7 +5,7 @@ model_weights/
 app/models/
 pip-wheel-metadata/
 lightning_logs/
-
+.vscode/
 
 # Test-tube
 test_tube_logs/
