+++ .markdownlint.yml
@@ -1,0 +1,2 @@
+MD013: false  # headers with the same names
+MD024: false  # line length
