+++ pytorch_lightning/callbacks/__init__.py
@@ -1,6 +1,11 @@
-from .pt_callbacks import EarlyStopping, ModelCheckpoint, GradientAccumulationScheduler
+from .base import Callback
+from .early_stopping import EarlyStopping
+from .model_checkpoint import ModelCheckpoint
+from .gradient_accumulation_scheduler import GradientAccumulationScheduler
+
 
 __all__ = [
+    'Callback',
     'EarlyStopping',
     'ModelCheckpoint',
     'GradientAccumulationScheduler',
